requirements.txt
streamlit
google-generativeai
python-dotenv
streamlit_webrtc
opencv-python-headless